+++ lib/auth/authOptions.ts
@@ -1,0 +1,58 @@
+import  { NextAuthOptions, DefaultSession, Session } from "next-auth";
+import GoogleProvider from "next-auth/providers/google";
+import GithubProvider from "next-auth/providers/github";
+import TwitterProvider from "next-auth/providers/twitter";
+import { JWT } from "next-auth/jwt";
+declare module "next-auth" {
+    interface Session {
+      user: {
+        id?: string;
+        avatar?: string;
+      } & DefaultSession["user"];
+    }
+    interface JWT {
+      id?: string;
+      avatar?: string;
+    }
+  }
+  
+  export const AuthOptions: NextAuthOptions = {
+    providers: [
+      GoogleProvider({
+        clientId: process.env.GOOGLE_CLIENT_ID as string,
+        clientSecret: process.env.GOOGLE_CLIENT_SECRET as string,
+      }),
+      GithubProvider({
+        clientId: process.env.GITHUB_ID as string,
+        clientSecret: process.env.GITHUB_SECRET as string,
+      }),
+      TwitterProvider({
+        clientId: process.env.TWITTER_CLIENT_ID as string,
+        clientSecret: process.env.TWITTER_CLIENT_SECRET as string,
+      }),
+    ],
+    session: {
+      strategy: "jwt",
+    },
+    callbacks: {
+      async jwt({ token, user }: { token: JWT; user?: any }): Promise<JWT> {
+        if (user) {
+          token.id = user.id;
+          token.avatar = user.image;
+        }
+        return token;
+      },
+      async session({ session, token }: { session: Session; token: JWT }) {
+        if (!session.user) {
+          session.user = { name: "", email: "", image: "" }; 
+        }
+        session.user.id = token.id as string;
+        session.user.avatar = token.avatar as string;
+        return session;
+      },
+    },
+    secret: process.env.NEXTAUTH_SECRET,
+    pages: {
+      signIn: "/login",
+    },
+  };
